+++ R/compare.R
@@ -6,12 +6,11 @@
 #' @param ... At least two objects returned by \code{\link{loo}} or
 #'   \code{\link{waic}}.
 #' @return A vector or matrix with class \code{'compare.loo'}. If \code{...}
-#'   contains more than two objects then a matrix is returned. This matrix
-#'   summarizes the objects and also reports model weights (the posterior
-#'   probability that each model has the best expected out-of-sample predictive
-#'   accuracy). If \code{...} contains exactly two objects then the difference
-#'   in expected predictive accuracy and the standard error of the difference
-#'   are returned (see Details) in addition to model weights.
+#'   contains more than two objects then a matrix of summary information is
+#'   returned. If \code{...} contains exactly two objects then the difference in
+#'   expected predictive accuracy and the standard error of the difference are
+#'   returned (see Details). The difference will be positive if the expected
+#'   predictive accuracy for the second model is higher.
 #'
 #' @details When comparing two fitted models, we can estimate the difference in
 #'   their expected predictive accuracy by the difference in \code{elpd_waic} or
@@ -26,6 +25,12 @@
 #'   approach of comparing differences of deviances to a Chi-squared
 #'   distribution, a practice derived for Gaussian linear models or
 #'   asymptotically, and which only applies to nested models in any case.
+#'
+#' @note In previous versions of \pkg{loo} model weights were also reported by
+#'   \code{compare}. We have removed the weights because they were based only on
+#'   the point estimate of the elpd values ignoring the uncertainty. We are
+#'   currently working on something similar to these weights that also accounts
+#'   for uncertainty, which will be included in future versions of \pkg{loo}.
 #'
 #' @template loo-paper-reference
 #'
@@ -43,9 +48,11 @@
 #'
 compare <- function(...) {
   dots <- list(...)
-  nms <- as.character(match.call())[-1L]
-  if (!all(sapply(dots, is.loo))) stop("All inputs should have class 'loo'.")
-  if (length(dots) <= 1L) stop("'compare' requires at least two models.")
+  nms <- as.character(match.call(expand.dots = TRUE))[-1L]
+  if (!all(sapply(dots, is.loo)))
+    stop("All inputs should have class 'loo'.", call. = FALSE)
+  if (length(dots) <= 1L)
+    stop("'compare' requires at least two models.", call. = FALSE)
   else if (length(dots) == 2L) {
     a <- dots[[1L]]
     b <- dots[[2L]]
@@ -59,28 +66,22 @@ compare <- function(...) {
     sqrtN <- sqrt(Na)
     elpd <- grep("^elpd", colnames(pa))
     diff <- pb[, elpd] - pa[, elpd]
-    uwts <- c(sum(pa[, elpd]), sum(pb[, elpd]))
-    uwts <- exp(uwts - max(uwts))
-    wts <- uwts / sum(uwts)
-    comp <- c(elpd_diff = sum(diff), se = sqrtN * sd(diff),
-              weight1 = wts[1L], weight2 = wts[2L])
+    comp <- c(elpd_diff = sum(diff), se = sqrtN * sd(diff))
     structure(comp, class = "compare.loo")
   }
   else {
     Ns <- sapply(dots, function(x) nrow(x$pointwise))
     if (!all(Ns == Ns[1L]))
-      stop("Not all models have the same number of data points.")
+      stop("Not all models have the same number of data points.", call. = FALSE)
     sel <- grep("pointwise|pareto_k", names(dots[[1L]]), invert = TRUE)
     x <- sapply(dots, function(x) unlist(x[sel]))
     colnames(x) <- nms
     rnms <- rownames(x)
-    uwts <- x[grep("^elpd", rnms), ]
-    uwts <- exp(uwts - max(uwts))
-    comp <- rbind(x, weights = uwts / sum(uwts))
-    col_ord <- order(uwts, decreasing = TRUE)
-    patts <- c("^waic$|^looic$", "^se_waic$|^se_looic$", "elpd", "p_", "weights")
+    comp <- x
+    patts <- c("^waic$|^looic$", "^se_waic$|^se_looic$", "elpd", "p_")
     row_ord <- unlist(sapply(patts, function(p) grep(p, rownames(comp))),
                       use.names = FALSE)
+    col_ord <- order(x[grep("^elpd", rnms), ], decreasing = TRUE)
     comp <- t(comp[row_ord, col_ord])
     class(comp) <- c("compare.loo", class(comp))
     comp