+++ R/extract_log_lik.R
@@ -30,20 +30,20 @@
 #'  \code{for (n in 1:N) log_lik[n] <- normal_log(y[n], X[n] * beta, sigma);}
 #'
 #' @references
-#' Stan Development Team (2015). Stan: A C++ library for probability and
-#' sampling, version 2.6. \url{mc-stan.org}.
+#' Stan Development Team (2016). Stan: A C++ library for probability and
+#' sampling, version 2.9. \url{http://mc-stan.org/documentation/}.
 #'
-#' Stan Development Team (2015). RStan, version 2.6.
-#' \url{mc-stan.org/rstan.html}.
+#' Stan Development Team (2016). RStan, version 2.9.
+#' \url{http://mc-stan.org/interfaces/rstan.html}.
 #'
 extract_log_lik <- function(stanfit, parameter_name = "log_lik") {
   if (!inherits(stanfit, "stanfit"))
     stop("Not a stanfit object.")
   if (stanfit@mode != 0)
     stop("Stan model does not contain posterior draws.")
-  log_lik <- try(as.matrix(stanfit, pars = parameter_name), silent = TRUE)
+  log_lik <- try(as.matrix(stanfit, pars = parameter_name), silent = TRUE) # nocov start
   if (inherits(log_lik, "try-error"))
     stop("Please load the 'rstan' package.")
   colnames(log_lik) <- NULL
   log_lik
-}
+} # nocov end